Add home parameter to user resource
The gitolite class provides the ability to use a custom home location
for the user under which gitolite will operate. However, that location
is not passed to the user resource embedded in the class, causing it to
use the system defaults for user home creation. In cases where /home is
unavailable or needs to be relocated, this can be confusing or
nonfunctional.

This commit ensures that the user resource respects the home property as
passed by the `home_path` parameter on the gitolite class.